requirements.txt
google-auth-oauthlib==0.4.6
google-auth==2.2.1
gspread==4.0.1
oauthlib==3.1.1
python-dotenv==0.19.0
vk-api==11.9.4
loguru~=0.5.3
requests~=2.26.0